src/requirements.txt
auto_mix_prep==0.2.0
boto3==1.34.93
botocore==1.34.93
docker_py==1.10.6
hydra-core==1.3.2
omegaconf==2.3.0
paramiko==3.4.0
